Annotated org.eclipse.kura.api 1.1.0
This is a version of the Kura API bundle released with Kura 2.1.0
annotated with OSGi R6 @ConsumerType and @ProviderType.
This version of the bundle is released with version 1.1.1
and the micro version of all the exports has been incremented.

The bnd-baseline-maven-plugin requires that the baseline
is properly annotated. See also:

https://github.com/bndtools/bnd/issues/1890

Signed-off-by: Cristiano De Alti <cristiano.dealti@eurotech.com>